requirements.txt
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.1
openpyxl==3.1.2
pandas==2.1.2
python-dateutil==2.8.2
tqdm==4.66.1
urllib3==2.0.7
cyvcf2==0.30.28
pymongo==4.6.1
pydantic==2.6.4
pydantic_core==2.16.3
typing==3.7.4.3
typing_extensions==4.11.0
ga4gh.vrs[extras]==2.1.3
seqrepo==0.0.0